dev_ns: adopt the android alarm-dev driver
Add device namespace logic to Android alarm-dev driver - so processes in a
namespace would only see alarms in its own namespace.

Place per-namespace data into struct alarm_dev_ns, and convert the driver
to operate on the current process's context.

The exception is an ioctl call with ANDROID_ALARM_SET_RTC, in which case
the fact that the RTC setting changes must be communicated to all of the
contexts (device namespaces) so it takes effect in each namespace. This
is by calling the new propagate_alarm_time_change().

  [ ported Cellrox code to Tegra drivers/rtc/alarm-dev.c - Jeremy ]

Changelog:
[16-Aug-2013] v1 - initial version

Change-Id: I82c0f5453d3f227a9b2323d823f884a251e8f98a
Signed-off-by: Oren Laadan <orenl@cellrox.com>
Signed-off-by: Amir Goldstein <amir@cellrox.com>
Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>