cells: fbearlysuspend virtualization
The framebuffer earlysuspend driver presents a sysfs interface to
userspace that allows processes to block until a particular framebuffer
state is valid. For example, a user space read from the
/sys/power/wait_for_fb_sleep file will block until the system is
sleeping and user space should stop drawing activities.

The virtualization of this driver isolates the framebuffer "sleep"
state using device namespaces. However, it is possible for the init
device namespace to wake up or put to sleep the currently active device
namespace.

Change-Id: Iae4cd8bb9276843c9ed02096c3664358b6a66d5e
Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>